requirements.txt
alembic
aniso8601
annotated-types
anyio
bcrypt
blinker
certifi
charset-normalizer
click
colorama
ecdsa
fastapi
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
greenlet
gunicorn
idna
itsdangerous
Jinja2
Mako
MarkupSafe
marshmallow
marshmallow-sqlalchemy
packaging
passlib
pip
psycopg2-binary
pyasn1
pydantic
pydantic_core
PyJWT
python-dotenv
python-jose
pytz
requests
rsa
setuptools
six
sniffio
SQLAlchemy
starlette
typing_extensions
urllib3
Werkzeug